implemented remaining feedback
